+++ src/pages/cart/index.tsx
@@ -2,18 +2,10 @@ import React, { useEffect, useReducer } from 'react';
 import { useRouter } from 'next/router';
 import { GetServerSideProps } from 'next';
 import DefaultContainerPage from 'src/components/common/DefaultContainer';
-import styled, { css } from 'styled-components';
 import { Buttons } from '@components';
 import CartItem from 'src/components/Cart/CartItem';
 import axios from 'axios';
-
-interface BarProps {
-  flex: number;
-}
-
-interface CheckBoxProps {
-  checked: boolean;
-}
+import * as Styled from './styled';
 
 export interface CartData {
   _id: number;
@@ -219,20 +211,20 @@ function CartPage({ data }: CartDataProps) {
 
   return (
     <DefaultContainerPage>
-      <PageTitle>장바구니</PageTitle>
-      <SectionBar>
-        <ContainerCheckBox>
-          <ContainerCheck
+      <Styled.PageTitle>장바구니</Styled.PageTitle>
+      <Styled.SectionBar>
+        <Styled.ContainerCheckBox>
+          <Styled.ContainerCheck
             checked={cartData
               .map((item: any) => item.checked)
               .every((item: boolean) => item === true)}
             onClick={checkedAll}
           />
-        </ContainerCheckBox>
-        <TextBar flex={55}>상품정보</TextBar>
-        <TextBar flex={15}>수량</TextBar>
-        <TextBar flex={25}>상품금액</TextBar>
-      </SectionBar>
+        </Styled.ContainerCheckBox>
+        <Styled.TextBar flex={55}>상품정보</Styled.TextBar>
+        <Styled.TextBar flex={15}>수량</Styled.TextBar>
+        <Styled.TextBar flex={25}>상품금액</Styled.TextBar>
+      </Styled.SectionBar>
 
       {cartData?.map((item: CartProps) => {
         return (
@@ -244,27 +236,27 @@ function CartPage({ data }: CartDataProps) {
         );
       })}
 
-      <SectionPrice>
-        <ContainerTextBox>
-          <TextPriceTitle>총 상품금액</TextPriceTitle>
-          <TextPrice>{totalprice}</TextPrice>
-        </ContainerTextBox>
-        <TextOperator>-</TextOperator>
-        <ContainerTextBox>
-          <TextPriceTitle>상품 할인</TextPriceTitle>
-          <TextPrice>0</TextPrice>
-        </ContainerTextBox>
-        <TextOperator>+</TextOperator>
-        <ContainerTextBox>
-          <TextPriceTitle>배송비</TextPriceTitle>
-          <TextPrice>{deliveryprice}</TextPrice>
-        </ContainerTextBox>
-        <ContainerTextBox>
-          <TextExpectTitle>결제 예정 금액</TextExpectTitle>
-          <TextExpectPrice>{finalprice}</TextExpectPrice>
-        </ContainerTextBox>
-      </SectionPrice>
-      <ConTainerBtn>
+      <Styled.SectionPrice>
+        <Styled.ContainerTextBox>
+          <Styled.TextPriceTitle>총 상품금액</Styled.TextPriceTitle>
+          <Styled.TextPrice>{totalprice}</Styled.TextPrice>
+        </Styled.ContainerTextBox>
+        <Styled.TextOperator>-</Styled.TextOperator>
+        <Styled.ContainerTextBox>
+          <Styled.TextPriceTitle>상품 할인</Styled.TextPriceTitle>
+          <Styled.TextPrice>0</Styled.TextPrice>
+        </Styled.ContainerTextBox>
+        <Styled.TextOperator>+</Styled.TextOperator>
+        <Styled.ContainerTextBox>
+          <Styled.TextPriceTitle>배송비</Styled.TextPriceTitle>
+          <Styled.TextPrice>{deliveryprice}</Styled.TextPrice>
+        </Styled.ContainerTextBox>
+        <Styled.ContainerTextBox>
+          <Styled.TextExpectTitle>결제 예정 금액</Styled.TextExpectTitle>
+          <Styled.TextExpectPrice>{finalprice}</Styled.TextExpectPrice>
+        </Styled.ContainerTextBox>
+      </Styled.SectionPrice>
+      <Styled.ConTainerBtn>
         <Buttons.Custom
           width={22}
           height={6.8}
@@ -275,7 +267,7 @@ function CartPage({ data }: CartDataProps) {
         >
           주문하기
         </Buttons.Custom>
-      </ConTainerBtn>
+      </Styled.ConTainerBtn>
     </DefaultContainerPage>
   );
 }
@@ -292,121 +284,4 @@ export const getServerSideProps: GetServerSideProps = async () => {
   };
 };
 
-const PageTitle = styled.h2`
-  font-size: 40px;
-  font-weight: bold;
-  text-align: center;
-  padding: 30px 0px;
-`;
-
-const SectionBar = styled.section`
-  display: flex;
-  padding: 22px 0px 22px;
-  margin-bottom: 35px;
-  border-radius: 10px;
-  background-color: #e5e5e5;
-  text-align: center;
-`;
-
-const ContainerCheckBox = styled.div`
-  flex-basis: 5%;
-`;
-
-const ContainerCheck = styled.div<CheckBoxProps>`
-  position: relative;
-  width: 20px;
-  height: 20px;
-  border: 2px solid #21bf48;
-  border-radius: 50%;
-  margin: 0 auto;
-
-  ${(props) =>
-    props.checked &&
-    css`
-      &::after {
-        content: '';
-        position: absolute;
-        top: 50%;
-        left: 50%;
-        transform: translate(-50%, -50%);
-        display: block;
-        width: 12px;
-        height: 12px;
-        background-color: #21bf48;
-        border-radius: 50%;
-      }
-    `}
-`;
-
-const TextBar = styled.p<BarProps>`
-  font-size: 18px;
-  flex-basis: ${(props) => props.flex}%;
-`;
-
-const SectionPrice = styled.section`
-  display: flex;
-  align-items: center;
-  border-radius: 10px;
-  padding: 46px 0;
-  margin-top: 80px;
-  background-color: #e5e5e5;
-`;
-
-const ContainerTextBox = styled.div`
-  width: 320px;
-  text-align: center;
-`;
-
-const TextPriceTitle = styled.p`
-  font-size: 16px;
-  margin-bottom: 18px;
-`;
-
-const TextPrice = styled.p`
-  font-size: 24px;
-  font-weight: bold;
-  &::after {
-    content: '원';
-    display: inline-block;
-    font-size: 16px;
-    font-weight: 400;
-    vertical-align: middle;
-  }
-`;
-
-const TextOperator = styled.span`
-  display: inline-block;
-  width: 34px;
-  height: 34px;
-  background-color: white;
-  border-radius: 50%;
-  color: #e5e5e5;
-  font-size: 35px;
-  text-align: center;
-`;
-
-const TextExpectTitle = styled.p`
-  font-size: 16px;
-  font-weight: bold;
-  margin-bottom: 10px;
-`;
-const TextExpectPrice = styled.p`
-  font-size: 36px;
-  font-weight: bold;
-  color: red;
-  &::after {
-    content: '원';
-    display: inline-block;
-    font-size: 18px;
-    font-weight: 400;
-    vertical-align: middle;
-  }
-`;
-
-const ConTainerBtn = styled.div`
-  display: flex;
-  justify-content: center;
-  align-content: center;
-  margin-top: 40px;
-`;
 export default CartPage;
